Toy 1 : Add DMDA-based PETSc toy
This attempts to solve problems 7.1 and 7.2 in Gerya's textbook,
in serial using a collection of DMDA objects. It is intended as a
reference for this test case (though it does NOT currently give
numerically-identical results to the MATLAB solutions) and to expose
issues with the multiple-DM approach.

Output is done with xdmf files which can be examined in Paraview.