Add Toy 2 : DMStag first prototype
Toy 2 solves the same problem as toy 1, namely exercises 7.1 and 7.2 in
Gerya's texbook.

It does this by defining a distributed staggered grid DM object for
PETSc and using several of these to define a solution, parameters, and
grids for output.

As with toy1, the solution does NOT precisely match the reference MATLAB
solution.

The prototype is only 2D.

Efficient computation of the local-to-global map is missing (as it would
be tedious rearrangement of the logic for computing the local-to-global
VecScatters) but this is required for the next implementation.