Tests: initial framework
Use pyTestHarness to define an initial test using the 2d demo.

To maintain the "independence" of the demo (which is mostly a miniapp
using the library, not a binary to usually be built with the library),
the separate makefile is maintained at the cost of an uglier process
building the binary to go into STAGBL_ARCH/bin.