StagBLArray, StagBLSystem: Provide most of the implementation to add a new "simple" implementation
These are currently only for one processor (though hopefully extensible
to more, with a parallel linear solver).

DMStag is still relied upon heavily.

The construction of the Stokes systems is made to rely less critically
on PETSc Mat and Vec datatypes.

To enable this, several new API functions are added.

StagBLGrid now contains information on both which kind of system and
which kind of array to create. These would almost always need to be
compatible.

Tests are updated to allow selection of the new implementations.